Table refreshing automatically now. 🎉
